requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.17.1
click==8.2.0
cryptography==44.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.103.1
greenlet==3.2.2
h11==0.16.0
httptools==0.6.4
idna==3.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.21
starlette==0.27.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.23.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
